seperate address components
