+++ src/components/pages/addresses/index.js
@@ -1,0 +1,4 @@
+export { default as Address } from "./Address";
+export { default as AddressLoader } from "./AddressLoader";
+export { default as AddressList } from "./AddressList";
+export { default as AddressListLoader } from "./AddressListLoader";
